+++ domain/src/main/java/com/ayanpandey/blink/domain/contract/DocumentRenderer.kt
@@ -1,0 +1,8 @@
+package com.ayanpandey.blink.domain.contract
+
+import com.ayanpandey.blink.domain.model.DocumentType
+
+interface DocumentRenderer {
+    val supportedType: DocumentType
+    val name: String
+}