+++ domain/src/main/java/com/ayanpandey/blink/domain/di/DomainContainer.kt
@@ -5,4 +5,5 @@ import com.ayanpandey.blink.domain.repository.FileResolver
 
 interface DomainContainer : AppContainer {
     val fileResolver: FileResolver
+    val documentViewer: com.ayanpandey.blink.domain.contract.DocumentViewer
 }